refactor(regional-guide): ViewModel 테스트를 흐름별 파일로 분리
- RegionalGuideViewModelTest를 selector/search/address/candidate/favorite 흐름별 파일로 분리
- ViewModel 생성 함수와 fake repository를 공통 fixture로 이동
- 기존 테스트 커버리지와 #152 주소 검색 테스트 유지
- 프로덕션 코드 변경 없이 테스트 구조만 정리

Refs #167
