+++ presentation/src/test/java/com/team/yeogibeoryeo/presentation/regionalguide/RegionalGuideFavoriteViewModelTest.kt
@@ -1,0 +1,257 @@
+﻿package com.team.yeogibeoryeo.presentation.regionalguide
+
+import com.team.yeogibeoryeo.domain.favorite.model.Favorite
+import com.team.yeogibeoryeo.domain.favorite.model.FavoriteTargetType
+import com.team.yeogibeoryeo.domain.favorite.model.RegionalGuideFavoriteKey
+import com.team.yeogibeoryeo.domain.favorite.model.RegionalGuideFavoriteSnapshot
+import com.team.yeogibeoryeo.domain.favorite.model.toFavoriteSnapshot
+import com.team.yeogibeoryeo.domain.favorite.repository.FavoriteRepository
+import com.team.yeogibeoryeo.domain.favorite.repository.RegionalGuideFavoriteRepository
+import com.team.yeogibeoryeo.domain.favorite.repository.RegionalGuideFavoriteSnapshotRepository
+import com.team.yeogibeoryeo.domain.favorite.usecase.GetRegionalGuideFavoriteSnapshotUseCase
+import com.team.yeogibeoryeo.domain.favorite.usecase.ObserveFavoriteUseCase
+import com.team.yeogibeoryeo.domain.favorite.usecase.ToggleRegionalGuideFavoriteUseCase
+import com.team.yeogibeoryeo.domain.region.model.Region
+import com.team.yeogibeoryeo.domain.region.repository.RegionOptionsRepository
+import com.team.yeogibeoryeo.domain.region.repository.RegionRepository
+import com.team.yeogibeoryeo.domain.region.usecase.ClassifyRegionSearchInputUseCase
+import com.team.yeogibeoryeo.domain.region.usecase.ExtractRegionFromAddressUseCase
+import com.team.yeogibeoryeo.domain.region.usecase.GetEupmyeondongOptionsUseCase
+import com.team.yeogibeoryeo.domain.region.usecase.GetSidoOptionsUseCase
+import com.team.yeogibeoryeo.domain.region.usecase.GetSigunguOptionsUseCase
+import com.team.yeogibeoryeo.domain.region.usecase.NormalizeRegionForRegionalGuideUseCase
+import com.team.yeogibeoryeo.domain.region.usecase.ResolveRegionFromKeywordUseCase
+import com.team.yeogibeoryeo.domain.regionalguide.model.RegionalDisposalGuide
+import com.team.yeogibeoryeo.domain.regionalguide.model.RegionalGuideQuery
+import com.team.yeogibeoryeo.domain.regionalguide.repository.RegionalDisposalGuideRepository
+import com.team.yeogibeoryeo.domain.regionalguide.usecase.GetRegionalDisposalGuideUseCase
+import com.team.yeogibeoryeo.domain.regionalguide.usecase.NormalizeRegionalGuideQueryUseCase
+import com.team.yeogibeoryeo.domain.regionalguide.usecase.SelectRegionalGuideCandidateUseCase
+import com.team.yeogibeoryeo.presentation.regionalguide.model.RegionSearchCandidateUiModel
+import kotlinx.coroutines.CompletableDeferred
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.map
+import kotlinx.coroutines.test.StandardTestDispatcher
+import kotlinx.coroutines.test.TestDispatcher
+import kotlinx.coroutines.test.advanceTimeBy
+import kotlinx.coroutines.test.advanceUntilIdle
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.runTest
+import kotlinx.coroutines.test.setMain
+import org.junit.Assert.assertEquals
+import org.junit.Assert.assertNull
+import org.junit.Assert.assertTrue
+import org.junit.Rule
+import org.junit.Test
+import org.junit.rules.TestWatcher
+import org.junit.runner.Description
+
+
+@OptIn(ExperimentalCoroutinesApi::class)
+class RegionalGuideFavoriteViewModelTest {
+
+    @get:Rule
+    val mainDispatcherRule = RegionalGuideMainDispatcherRule()
+
+    @Test
+    fun `favorite click stores regional guide favorite and observes state`() = runTest {
+        val guide =
+            RegionalDisposalGuide(
+                region = Region(sido = "Sido", sigungu = "Sigungu"),
+                targetRegionName = "Zone A",
+                managementZoneName = "Management A",
+                schedules = emptyList(),
+            )
+        val favoriteRepository = FakeFavoriteRepository()
+        val snapshotRepository = FakeRegionalGuideFavoriteSnapshotRepository()
+        val viewModel =
+            createViewModel(
+                regionRepository = FakeRegionRepository(resolvedRegion = guide.region),
+                regionalGuideRepository = FakeRegionalDisposalGuideRepository(candidates = listOf(guide)),
+                favoriteRepository = favoriteRepository,
+                regionalGuideSnapshotRepository = snapshotRepository,
+                regionalGuideFavoriteRepository =
+                    FakeRegionalGuideFavoriteRepository(
+                        favoriteRepository = favoriteRepository,
+                        snapshotRepository = snapshotRepository,
+                    ),
+            )
+        advanceUntilIdle()
+
+        viewModel.onSearchKeywordChanged("Sigungu")
+        viewModel.searchCurrentKeyword()
+        advanceUntilIdle()
+
+        assertEquals(false, (viewModel.uiState.value as RegionalGuideUiState.Success).isFavorite)
+
+        viewModel.onFavoriteClick()
+        advanceUntilIdle()
+
+        val snapshot = guide.toFavoriteSnapshot()
+        assertEquals(true, (viewModel.uiState.value as RegionalGuideUiState.Success).isFavorite)
+        assertEquals(true, favoriteRepository.isFavorite(FavoriteTargetType.REGIONAL_GUIDE, snapshot.targetId))
+        assertEquals(snapshot, snapshotRepository.getSnapshot(snapshot.targetId))
+    }
+
+    @Test
+    fun `favorite target id restores saved regional guide candidate`() = runTest {
+        val region = Region(sido = "Sido", sigungu = "Sigungu")
+        val firstGuide =
+            RegionalDisposalGuide(
+                region = region,
+                targetRegionName = "Zone A",
+                schedules = emptyList(),
+            )
+        val savedGuide =
+            RegionalDisposalGuide(
+                region = region,
+                targetRegionName = "Zone B",
+                managementZoneName = "Management B",
+                schedules = emptyList(),
+            )
+        val snapshot = savedGuide.toFavoriteSnapshot()
+        val favoriteRepository =
+            FakeFavoriteRepository(
+                initialFavorites =
+                    listOf(
+                        Favorite(
+                            type = FavoriteTargetType.REGIONAL_GUIDE,
+                            targetId = snapshot.targetId,
+                            savedAtMillis = 1L,
+                        ),
+                    ),
+            )
+        val viewModel =
+            createViewModel(
+                regionalGuideRepository =
+                    FakeRegionalDisposalGuideRepository(candidates = listOf(firstGuide, savedGuide)),
+                favoriteRepository = favoriteRepository,
+                regionalGuideSnapshotRepository =
+                    FakeRegionalGuideFavoriteSnapshotRepository(snapshots = listOf(snapshot)),
+            )
+        advanceUntilIdle()
+
+        viewModel.loadByFavoriteTargetId(snapshot.targetId)
+        advanceUntilIdle()
+
+        val state = viewModel.uiState.value as RegionalGuideUiState.Success
+        assertEquals("Zone B", state.guide.targetRegionName)
+        assertEquals(true, state.isFavorite)
+    }
+
+    @Test
+    fun `favorite target id restores candidate with same target region and different management zone`() = runTest {
+        val region = Region(sido = "대전광역시", sigungu = "유성구", eupmyeondong = "반석동")
+        val firstGuide =
+            RegionalDisposalGuide(
+                region = region,
+                targetRegionName = "반석동 일부지역",
+                managementZoneName = "노은2동",
+                schedules = emptyList(),
+            )
+        val savedGuide =
+            RegionalDisposalGuide(
+                region = region,
+                targetRegionName = "반석동 일부지역",
+                managementZoneName = "노은3동",
+                schedules = emptyList(),
+            )
+        val snapshot = savedGuide.toFavoriteSnapshot()
+        val favoriteRepository =
+            FakeFavoriteRepository(
+                initialFavorites =
+                    listOf(
+                        Favorite(
+                            type = FavoriteTargetType.REGIONAL_GUIDE,
+                            targetId = snapshot.targetId,
+                            savedAtMillis = 1L,
+                        ),
+                    ),
+            )
+        val viewModel =
+            createViewModel(
+                regionalGuideRepository =
+                    FakeRegionalDisposalGuideRepository(candidates = listOf(firstGuide, savedGuide)),
+                favoriteRepository = favoriteRepository,
+                regionalGuideSnapshotRepository =
+                    FakeRegionalGuideFavoriteSnapshotRepository(snapshots = listOf(snapshot)),
+            )
+        advanceUntilIdle()
+
+        viewModel.loadByFavoriteTargetId(snapshot.targetId)
+        advanceUntilIdle()
+
+        val state = viewModel.uiState.value as RegionalGuideUiState.Success
+        assertEquals("반석동 일부지역", state.guide.targetRegionName)
+        assertEquals("노은3동", state.guide.managementZoneName)
+        assertEquals(true, state.isFavorite)
+    }
+
+    @Test
+    fun `legacy regional guide favorite key is observed and removed with current snapshot`() = runTest {
+        val region = Region(sido = "Sido", sigungu = "Sigungu", eupmyeondong = "Dong")
+        val guide =
+            RegionalDisposalGuide(
+                region = region,
+                targetRegionName = "Target",
+                managementZoneName = "Management",
+                schedules = emptyList(),
+            )
+        val currentSnapshot = guide.toFavoriteSnapshot()
+        val legacyTargetId =
+            RegionalGuideFavoriteKey(
+                sido = region.sido,
+                sigungu = region.sigungu,
+                eupmyeondong = region.eupmyeondong,
+                targetRegionName = guide.targetRegionName,
+            ).encodeLegacy()
+        val legacySnapshot = currentSnapshot.copy(targetId = legacyTargetId, managementZoneName = null)
+        val favoriteRepository =
+            FakeFavoriteRepository(
+                initialFavorites =
+                    listOf(
+                        Favorite(
+                            type = FavoriteTargetType.REGIONAL_GUIDE,
+                            targetId = legacyTargetId,
+                            savedAtMillis = 1L,
+                        ),
+                    ),
+            )
+        val snapshotRepository =
+            FakeRegionalGuideFavoriteSnapshotRepository(snapshots = listOf(legacySnapshot))
+        val viewModel =
+            createViewModel(
+                regionRepository = FakeRegionRepository(resolvedRegion = region),
+                regionalGuideRepository = FakeRegionalDisposalGuideRepository(candidates = listOf(guide)),
+                favoriteRepository = favoriteRepository,
+                regionalGuideSnapshotRepository = snapshotRepository,
+                regionalGuideFavoriteRepository =
+                    FakeRegionalGuideFavoriteRepository(
+                        favoriteRepository = favoriteRepository,
+                        snapshotRepository = snapshotRepository,
+                    ),
+            )
+        advanceUntilIdle()
+
+        viewModel.onSearchKeywordChanged("Sigungu")
+        viewModel.searchCurrentKeyword()
+        advanceUntilIdle()
+
+        assertEquals(true, (viewModel.uiState.value as RegionalGuideUiState.Success).isFavorite)
+
+        viewModel.onFavoriteClick()
+        advanceUntilIdle()
+
+        assertEquals(false, (viewModel.uiState.value as RegionalGuideUiState.Success).isFavorite)
+        assertEquals(false, favoriteRepository.isFavorite(FavoriteTargetType.REGIONAL_GUIDE, legacyTargetId))
+        assertEquals(false, favoriteRepository.isFavorite(FavoriteTargetType.REGIONAL_GUIDE, currentSnapshot.targetId))
+        assertNull(snapshotRepository.getSnapshot(legacyTargetId))
+        assertNull(snapshotRepository.getSnapshot(currentSnapshot.targetId))
+    }
+
+}
+
+
